Extract #kind_of? on singletons to ClassRelationships module
